+++ custom_components/nidia_magic_composer/const.py
@@ -7,7 +7,7 @@ DOMAIN = "nidia_magic_composer"
 NAME = "Nidia Magic Composer"
 
 # Version
-VERSION = "0.1.9"
+VERSION = "0.2.0"
 
 # Panel configuration
 PANEL_NAME = "nidia-magic-composer"